- fixed bug 5205 (http://bugzilla.osafoundation.org/show_bug.cgi?id=5205)
git-svn-id: http://svn.osafoundation.org/chandler/trunk@9559 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/repository/persistence/DBRepository.py
@@ -508,7 +508,8 @@ class DBRepository(OnDemandRepository):
             if ramdb:
                 self._status &= ~Repository.RAMDB
             elif self._openFile is not None:
-                os.remove(self._openFile)
+                if os.path.exists(self._openFile):
+                    os.remove(self._openFile)
                 self._openFile = None
 
             self._status |= Repository.CLOSED
